+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## UNRELEASED
 
+* Added `timeout_seconds` (default value is 10) param to `LaterPayClient`. Backend API requests will time out now according to that value.
+
+
 ## 3.2.1
 
 * Standardised the usage of the (internal) `product` API (it was usable via path or parameter, and we're deprecating the path usage)